Student Database with GUI (but not finished)
This is another version. I made a GUI which is made using netbeans and eclipse. There's a lot of comments generated from the netbeans GUI builder and I forgot to ommit all of them.
All the action from button and textfield and etc are generally made from scratch so I can use custom actions like accessing and editing infos from the db.