Student Database with GUI (Final)
This is another version. I made a GUI which is made using netbeans and eclipse. There's a lot of comments generated from the netbeans GUI builder and I forgot to ommit all of them.
All the action from button and textfield and etc are generally made from scratch so I can use custom actions like accessing and editing infos from the db.
There is an error checking function for id numbers existing or not. I have not put any handling errors for invalid texts and data. I'm still new to action performed in GUI's. But still, this is a working program